Add link to sign up for newsletter after threshold
In order to make this work, the app needs to know about the number of
completed pomodoros on load, not just when one is completed. Thus, added
reading the cookie on mount and modifying state accordingly.

This led to complacations because other parts of the app were
monitoring for changes to the completed pomodoros in state. Fixed this
by adding a value that tracks completed pomodoros this session. This has
the benefit of addressing the bugs introduced by modifying the completed
pomodoros and keeps things in the store.