Make transition from old to new Pomodoro smooth
Before the behavior of the reset button varied from context to context.
If the timer was running, it would reset to 25 and stop. If the Pomodoro
was done, it would immediately start a new one.

Now makes it so that it resets to the starting time (25 min) and waits
for user to press play.

Downside: now it takes two clicks to start running a new Pomodoro. Not
sure if better or worse, may want to revisit.